refactor: 버튼 props 상수 파일 분리
