+++ src/components/common/button/index.js
@@ -1,4 +1,4 @@
 export { default as Button } from "./Button";
-export { default as GoBackButton } from "./GoBackbutton";
+export { default as GoBackButton } from "./GoBackButton";
 export { default as ExitButton } from "./ExitButton";
 export { default as SubmitButton } from "./SubmitButton";
