+++ src/constants/buttons/authPageButton.js
@@ -1,0 +1,13 @@
+export const loginButtonProps = {
+  type: "submit",
+  name: "login",
+  text: "로그인",
+  isDisabled: false,
+};
+
+export const joinButtonProps = {
+  type: "submit",
+  name: "join",
+  text: "가입하기",
+  isDisabled: true,
+};
