+++ src/constants/buttons/mytownPageButton.js
@@ -5,3 +5,11 @@ export const addressSearchOpenButtonProps = {
   img: { src: "/img/plus.svg", alt: "동네 추가" },
   isDisabled: false,
 };
+
+export const FindByCurrentLocationButtonProps = {
+  type: "submit",
+  name: "submit",
+  text: "현재 위치로 찾기",
+  img: { src: "/img/plus.svg", alt: "위치" },
+  isDisabled: false,
+};
