+++ src/constants/index.js
@@ -1,3 +1,3 @@
 export * from "./textField";
-export * from "./button";
+export * from "./buttons";
 export * from "./validateRule";
